requirements.txt
blinker==1.9.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.3.1
colorama==0.4.6
Flask==3.1.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.1.4
